ASL_to_English/requirements_api.txt
# Requirements for FastAPI server only
# (MediaPipe dependencies are in requirements_signtalk.txt)

# For FastAPI server
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6  # Required for file uploads (UploadFile)

# For environment variables
python-dotenv>=1.0.0

# For ElevenLabs API
elevenlabs>=0.1.0

# for Gemini API
google-genai>=0.3.0


